+++ src/main/java/ua/slavik/carwash/model/Employee.java
@@ -19,7 +19,7 @@ public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int age;
+    private Long age;
     private String firstName;
     private String lastName;
     private String phoneNumber;
